refactor: reorganize and consolidate wallet DTOs
This commit refactors the wallet data layer by consolidating DTOs into a centralized package and introducing domain mapping logic.

- **WalletsItem.kt**:
    - Relocated to `id.my.sendiko.fintrack.wallet.core.data.dto`.
    - Added a `toDomain()` extension function to map the DTO to the `Wallet` domain model.
- **PostWalletItem.kt**:
    - Created a new DTO specifically for wallet creation responses.
    - Included a `toDomain()` function for domain model transformation.
- **PostWalletResponse.kt & PostWalletRequest.kt**:
    - Moved from `wallet.create.data` to `wallet.core.data.dto.post`.
    - Updated `PostWalletResponse` to use the new `PostWalletItem`.
- **Cleanup**:
    - Removed redundant `Wallet` and `WalletsItem` classes from legacy packages to simplify the data structure.
