+++ shared/src/commonMain/kotlin/id/my/sendiko/fintrack/wallet/core/data/dto/WalletsItem.kt
@@ -1,0 +1,50 @@
+package id.my.sendiko.fintrack.wallet.core.data.dto
+
+import id.my.sendiko.fintrack.wallet.core.data.dto.getdetails.TransactionsItem
+import id.my.sendiko.fintrack.wallet.core.domain.Wallet
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class WalletsItem(
+
+    @SerialName("createdAt")
+	val createdAt: String,
+
+    @SerialName("balance")
+	val balance: Int,
+
+    @SerialName("purpose")
+	val purpose: String,
+
+    @SerialName("name")
+	val name: String,
+
+    @SerialName("id")
+	val id: String,
+
+    @SerialName("type")
+	val type: String,
+
+    @SerialName("transactions")
+	val transactions: List<TransactionsItem>,
+
+    @SerialName("userId")
+	val userId: String,
+
+    @SerialName("walletNumber")
+	val walletNumber: String?,
+
+    @SerialName("updatedAt")
+	val updatedAt: String
+) {
+
+	fun toDomain() = Wallet(
+        id = id,
+        name = name,
+        purpose = purpose,
+        type = type,
+        amount = balance.toDouble(),
+        number = walletNumber ?: ""
+    )
+}
